requirements.txt
psycopg2-binary~=2.9.9
wheel
setuptools
autobahn[twisted]~=23.1.2
twisted~=24.7.0
pympler
jsmin
psutil~=6.0.0
git+https://github.com/rossengeorgiev/aprs-python
aprslib~=0.7.2